@@ -1,7 +1,7 @@
 {
     "name": "website",
     "type": "module",
-    "version": "0.3",
+    "version": "0.4",
     "packageManager": "pnpm@9.4.0",
     "scripts": {
         "stamp": "echo \"COMMIT_HASH=$(git rev-parse --short HEAD)\"> .env && echo \"WEBSITE_VERSION=$(node -p \"require('./package.json').version\")\">>.env",
@@ -14,40 +14,41 @@
         "lint": "prettier --check . && eslint . && svelte-check --tsconfig ./tsconfig.json"
     },
     "dependencies": {
-        "@astrojs/mdx": "^3.1.7",
-        "@astrojs/svelte": "^5.7.1",
-        "@astrojs/tailwind": "^5.1.1",
-        "@astrojs/ts-plugin": "^1.10.2",
-        "astro": "^4.15.8",
-        "daisyui": "^4.12.10",
-        "fuse.js": "^7.0.0",
-        "katex": "^0.16.11",
-        "rehype-katex": "^7.0.1",
-        "remark-math": "^6.0.0",
-        "svelte": "^4.2.19",
-        "tailwindcss": "^3.4.12",
-        "theme-change": "^2.5.0"
+        "@fontsource/crimson-text": "^5.2.7",
+        "@fontsource/fira-code": "^5.2.7",
+        "katex": "^0.16.28"
     },
     "devDependencies": {
-        "@iconify-json/carbon": "^1.2.1",
-        "@iconify-json/line-md": "^1.2.1",
-        "@iconify-json/octicon": "^1.2.0",
-        "@tailwindcss/typography": "^0.5.15",
+        "@astrojs/mdx": "^4.3.13",
+        "@astrojs/sitemap": "^3.7.0",
+        "@astrojs/svelte": "^7.2.5",
+        "@astrojs/ts-plugin": "^1.10.6",
+        "@iconify-json/carbon": "^1.2.18",
+        "@iconify-json/line-md": "^1.2.11",
+        "@iconify-json/material-symbols": "^1.2.55",
+        "@iconify-json/octicon": "^1.2.20",
+        "@tailwindcss/typography": "^0.5.19",
+        "@tailwindcss/vite": "^4.1.18",
         "@types/eslint": "^9.6.1",
-        "autoprefixer": "^10.4.20",
-        "eslint": "^9.11.0",
-        "eslint-config-prettier": "^9.1.0",
-        "eslint-plugin-svelte": "^2.44.0",
-        "globals": "15.9.0",
-        "postcss": "^8.4.47",
-        "prettier": "^3.3.3",
+        "@types/hast": "^3.0.4",
+        "astro": "^5.17.1",
+        "eslint": "^9.39.2",
+        "eslint-config-prettier": "^10.1.8",
+        "eslint-plugin-svelte": "^3.15.0",
+        "globals": "16.0.0",
+        "prettier": "^3.8.1",
         "prettier-plugin-astro": "^0.14.1",
-        "prettier-plugin-svelte": "^3.2.6",
+        "prettier-plugin-svelte": "^3.4.1",
+        "rehype-katex": "^7.0.1",
+        "remark-math": "^6.0.0",
         "sharp": "^0.33.5",
-        "svelte-check": "^3.8.6",
-        "typescript": "^5.6.2",
-        "typescript-eslint": "^8.6.0",
-        "unplugin-icons": "^0.19.3",
-        "vite": "^5.4.7"
+        "svelte": "^5.50.1",
+        "svelte-check": "^4.3.6",
+        "tailwindcss": "^4.1.18",
+        "typescript": "^5.9.3",
+        "typescript-eslint": "^8.55.0",
+        "unist-util-visit": "^5.1.0",
+        "unplugin-icons": "^22.5.0",
+        "vite": "^6.4.1"
     }
 }
