@@ -1,22 +1,43 @@
 import { defineConfig } from "astro/config";
 import svelte from "@astrojs/svelte";
-import tailwind from "@astrojs/tailwind";
 import Icons from "unplugin-icons/vite";
 import mdx from "@astrojs/mdx";
 import remarkMath from "remark-math";
 import rehypeKatex from "rehype-katex";
+import rehypeCaptionImages from "./src/utils/rehype-caption-images.ts";
+import rehypeCaptionTables from "./src/utils/rehype-caption-tables.ts";
+import tailwindcss from "@tailwindcss/vite";
+import sitemap from "@astrojs/sitemap";
 
 // https://astro.build/config
 export default defineConfig({
-    integrations: [svelte(), tailwind(), mdx()],
+    integrations: [svelte(), mdx(), sitemap()],
     site: "https://bustosalex1.github.io",
     markdown: {
+        shikiConfig: {
+            themes: {
+                light: "catppuccin-latte",
+                dark: "dracula",
+            },
+        },
         remarkPlugins: [remarkMath],
         rehypePlugins: [
             [
                 rehypeKatex,
                 {
                     // katex plugin options
+                },
+            ],
+            [
+                rehypeCaptionImages,
+                {
+                    // rehype caption images options
+                },
+            ],
+            [
+                rehypeCaptionTables,
+                {
+                    // rehype caption tables options
                 },
             ],
         ],
@@ -31,6 +52,7 @@ export default defineConfig({
                 compiler: "astro",
                 autoInstall: true,
             }),
+            tailwindcss(),
         ],
     },
 });
