@@ -20,10 +20,14 @@ export default [
         },
     },
     {
-        files: ["**/*.svelte"],
+        files: ["**/*.svelte", "**/*.svelte.ts"],
         languageOptions: {
             parserOptions: {
                 parser: ts.parser,
+                extraFileExtensions: [".svelte", ".svelte.ts"],
+                svelteFeatures: {
+                    experimentalGenerics: true,
+                },
             },
         },
     },
@@ -35,7 +39,6 @@ export default [
             "astro.config.mjs",
             "src/env.d.ts",
             "tailwind.config.cjs",
-            "postcss.config.cjs",
         ],
     },
 ];
